requirements.txt
djangorestframework
djoser
django-filter
drf-yasg
isort
python-dotenv